+++ config/routes.rb
@@ -353,6 +353,7 @@ Rails.application.routes.draw do
       member do
         post "create_variation", to: "board_images#create_image_variation"
         post "create_edit", to: "board_images#create_image_edit"
+        post "create_text_image", to: "board_images#create_text_image"
         post "upload_audio", to: "board_images#upload_audio"
         post "reset_audio", to: "board_images#reset_audio"
         post "set_current_audio", to: "board_images#set_current_audio"